app/requirements.txt
chainlit==1.0.301
chromadb==0.4.22
langchain==0.1.8
python-dotenv==1.0.1
tiktoken==0.6.0
openai==1.12.0
langchain-community==0.0.24
langchain-openai==0.0.7
